Add YOLOv4 runner
