+++ models/demos/wormhole/yolov4/test_yolov4_performant.py
@@ -2,14 +2,15 @@
 
 # SPDX-License-Identifier: Apache-2.0
 
-import pytest
-import ttnn
 import time
+
+import pytest
 import torch
 from loguru import logger
 
+import ttnn
+from models.demos.yolov4.runner.performant_runner import YOLOv4PerformantRunner
 from models.utility_functions import run_for_wormhole_b0
-from models.demos.yolov4.tests.yolov4_perfomant_webdemo import Yolov4Trace2CQ
 
 
 @run_for_wormhole_b0()
@@ -38,9 +39,7 @@ def test_run_yolov4_trace_2cqs_inference(
     model_location_generator,
     resolution,
 ):
-    yolov4_trac2_2cq = Yolov4Trace2CQ()
-
-    yolov4_trac2_2cq.initialize_yolov4_trace_2cqs_inference(
+    performant_runner = YOLOv4PerformantRunner(
         device,
         batch_size,
         act_dtype,
@@ -49,25 +48,19 @@ def test_run_yolov4_trace_2cqs_inference(
         model_location_generator=None,
     )
 
-    inference_iter_count = 10
-    inference_time_iter = []
-    for iter in range(0, inference_iter_count):
-        input_shape = (1, 3, *resolution)
-
+    inference_times = []
+    for _ in range(10):
+        input_shape = (1, *resolution, 3)
         torch_input_tensor = torch.randn(input_shape, dtype=torch.float32)
-        n, c, h, w = torch_input_tensor.shape
-        torch_input_tensor = torch_input_tensor.permute(0, 2, 3, 1)
-        torch_input_tensor = torch_input_tensor.reshape(1, 1, h * w * n, c)
-        tt_inputs_host = ttnn.from_torch(torch_input_tensor, dtype=ttnn.bfloat16, layout=ttnn.ROW_MAJOR_LAYOUT)
-        tt_inputs_host = ttnn.pad(tt_inputs_host, [1, 1, n * h * w, 16], [0, 0, 0, 0], 0)
 
         t0 = time.time()
-        output = yolov4_trac2_2cq.execute_yolov4_trace_2cqs_inference(tt_inputs_host)
+        _ = performant_runner.run(torch_input_tensor)
         t1 = time.time()
-        inference_time_iter.append(t1 - t0)
-    yolov4_trac2_2cq.release_yolov4_trace_2cqs_inference()
-    inference_time_avg = round(sum(inference_time_iter) / len(inference_time_iter), 6)
-    # print(batch_size/inference_time_avg)
+        inference_times.append(t1 - t0)
+
+    performant_runner.release()
+
+    inference_time_avg = round(sum(inference_times) / len(inference_times), 6)
     logger.info(
         f"ttnn_yolov4_batch_size: {batch_size}, resolution: {resolution}. One inference iteration time (sec): {inference_time_avg}, FPS: {round(batch_size/inference_time_avg)}"
     )
