+++ models/demos/yolov4/resources/coco.names
