+++ models/demos/yolov4/resources/giraffe_320.jpg
