+++ models/experimental/yolo_evaluation/yolo_evaluation_utils.py
@@ -2,14 +2,15 @@
 
 # SPDX-License-Identifier: Apache-2.0
 
-import os
-import cv2
 import math
+import os
 import time
+from pathlib import Path
+
+import cv2
+import numpy as np
 import torch
 import torchvision
-import numpy as np
-from pathlib import Path
 from loguru import logger
