+++ tests/nightly/single_card/yolov4/test_yolov4_performant.py
@@ -1,0 +1,1 @@
+../../../../models/demos/wormhole/yolov4/test_yolov4_performant.py
