--- tests/nightly/single_card/yolov4/test_yolov4_performant_webdemo.py
@@ -1,1 +1,0 @@
-../../../../models/demos/wormhole/yolov4/test_yolov4_performant_webdemo.py
