+++ tests/ttnn/integration_tests/yolov4/test_ttnn_downsample1.py
@@ -2,17 +2,18 @@
 
 # SPDX-License-Identifier: Apache-2.0
 
-import ttnn
+import time
+
+import pytest
 import torch
-from models.utility_functions import skip_for_grayskull
-from models.demos.yolov4.reference.downsample1 import DownSample1
+from loguru import logger
+
+import ttnn
+from models.demos.yolov4.common import load_torch_model
 from models.demos.yolov4.ttnn.downsample1 import Down1
 from models.demos.yolov4.ttnn.model_preprocessing import create_ds1_model_parameters
+from models.utility_functions import skip_for_grayskull
 from tests.ttnn.utils_for_testing import assert_with_pcc
-import pytest
-import time
-from loguru import logger
-import os
 
 
 @skip_for_grayskull()
@@ -26,26 +27,9 @@ import os
 )
 def test_down1(device, reset_seeds, model_location_generator, resolution):
     torch.manual_seed(0)
-    model_path = model_location_generator("models", model_subdir="Yolo")
 
-    if model_path == "models":
-        if not os.path.exists("tests/ttnn/integration_tests/yolov4/yolov4.pth"):  # check if yolov4.th is availble
-            os.system(
-                "tests/ttnn/integration_tests/yolov4/yolov4_weights_download.sh"
-            )  # execute the yolov4_weights_download.sh file
-
-        weights_pth = "tests/ttnn/integration_tests/yolov4/yolov4.pth"
-    else:
-        weights_pth = str(model_path / "yolov4.pth")
-
+    torch_model = load_torch_model(model_location_generator, "down1")
     torch_input = torch.randn((1, 3, *resolution), dtype=torch.float)
-    torch_model = DownSample1()
-
-    torch_dict = torch.load(weights_pth)
-    ds_state_dict = {k: v for k, v in torch_dict.items() if (k.startswith("down1."))}
-    new_state_dict = dict(zip(torch_model.state_dict().keys(), ds_state_dict.values()))
-    torch_model.load_state_dict(new_state_dict)
-    torch_model.eval()
     ref = torch_model(torch_input)
 
     parameters = create_ds1_model_parameters(torch_model, torch_input, resolution, device)
@@ -56,7 +40,7 @@ def test_down1(device, reset_seeds, model_location_generator, resolution):
     ttnn_input = ttnn.from_torch(torch_input, dtype=ttnn.bfloat16)
 
     result_ttnn = ttnn_model(ttnn_input)
-    if resolution[0] == 320:
+    if resolution == (320, 320):
         start_time = time.time()
         for x in range(100):
             result_ttnn = ttnn_model(ttnn_input)
